+++ packages/framework/jest.config.js
@@ -1,0 +1,7 @@
+module.exports = {
+  preset: 'ts-jest',
+  testEnvironment: 'node',
+  testPathIgnorePatterns: [
+    "<rootDir>/src/__tests__/utils"
+  ]
+}